refactor: optional module naming cleanup
