+++ rigflow-server/src/waterfall/mod.rs
@@ -1,1 +1,1 @@
-pub mod simple;
+pub mod generator;
